Added tested code.
Deleted the old framework and added in the code which I wrote and tested for the parts I already have. So far, there is an arduino, the OLED screen and two force sensing resistors. The code currently displays the adafruit symbol, Sen. Proj. 2018, and then the net values from the force sensors. I have not yet confirmed what units the force sensors display.